utils/text_extraction.py
import os
from pdfminer.high_level import extract_text as extract_pdf_text
from docx import Document

def extract_text(file_path):
    ext = os.path.splitext(file_path)[1].lower()
    if ext == '.pdf':
        return extract_pdf_text(file_path)
    elif ext == '.docx':
        doc = Document(file_path)
        return '\n'.join([p.text for p in doc.paragraphs])
    else:
        raise ValueError('Unsupported file type: ' + ext)
