requirements.txt
fastapi
uvicorn
python-multipart
pdfminer.six
python-docx
spacy
scikit-learn
sentence-transformers
numpy
